+++ src/Addons/34thPRC_AceArsenalExtended/data/aax_models/vehicles/armour_standard/config.hpp
@@ -1,2 +1,3 @@
 #include "odst\config.hpp"
+#include "odst2\config.hpp"
 #include "mkvb\config.hpp"