+++ src/Addons/34thPRC_ArmourStandard/data/odst2/arid/config_vehicles.hpp
@@ -1,0 +1,60 @@
+class 34thPRC_ArmourStandard_ODST_M52_Rucksack_Arid : 34thPRC_ArmourStandard_ODST_M52_Rucksack_Urban
+{
+	displayName="[34th] M52 ODST Rucksack Arid";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\odst2\arid\34th_odst_rucksack_arid.paa",
+		"MA_Armor\data\Backpacks\M52_Rucksack\Straps_CO.paa"
+	};
+	class XtdGearInfo
+	{
+		model="34thPRC_ArmourStandard_ODST_M52_Rucksack_Urban";
+		variant="m52";
+		camo="arid";
+	};
+};
+class 34thPRC_ArmourStandard_ODST_M52_Rucksack_Medic_Arid : 34thPRC_ArmourStandard_ODST_M52_Rucksack_Medic_Urban
+{
+	displayName="[34th] M52-C ODST Rucksack Arid";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\odst2\arid\34th_odst_rucksack_medic_arid.paa",
+		"MA_Armor\data\Backpacks\M52_Rucksack\Straps_CO.paa"
+	};
+	class XtdGearInfo
+	{
+		model="34thPRC_ArmourStandard_ODST_M52_Rucksack_Urban";
+		variant="c";
+		camo="arid";
+	};
+};
+class 34thPRC_ArmourStandard_ODST_M52_Rucksack_Gunbag_Arid : 34thPRC_ArmourStandard_ODST_M52_Rucksack_Gunbag_Urban
+{
+	displayName="[34th] M52-G ODST Rucksack Arid";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\odst2\arid\34th_odst_rucksack_arid.paa",
+		"MA_Armor\data\Backpacks\M52_Rucksack\Straps_CO.paa"
+	};
+	class XtdGearInfo
+	{
+		model="34thPRC_ArmourStandard_ODST_M52_Rucksack_Urban";
+		variant="g";
+		camo="arid";
+	};
+};
+class 34thPRC_ArmourStandard_ODST_M52_Rucksack_Radio_Arid : 34thPRC_ArmourStandard_ODST_M52_Rucksack_Radio_Urban
+{
+	displayName="[34th] M52-R ODST Rucksack Arid";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\odst2\arid\34th_odst_rucksack_arid.paa",
+		"MA_Armor\data\Backpacks\M52_Rucksack\Straps_CO.paa"
+	};
+	class XtdGearInfo
+	{
+		model="34thPRC_ArmourStandard_ODST_M52_Rucksack_Urban";
+		variant="r";
+		camo="arid";
+	};
+};