+++ src/Addons/34thPRC_ArmourStandard/data/odst2/tropic/config_vehicles.hpp
@@ -1,0 +1,60 @@
+class 34thPRC_ArmourStandard_ODST_M52_Rucksack_Tropic : 34thPRC_ArmourStandard_ODST_M52_Rucksack_Urban
+{
+	displayName="[34th] M52 ODST Rucksack Tropic";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\odst2\tropic\34th_odst_rucksack_tropic.paa",
+		"MA_Armor\data\Backpacks\M52_Rucksack\Straps_CO.paa"
+	};
+	class XtdGearInfo
+	{
+		model="34thPRC_ArmourStandard_ODST_M52_Rucksack_Urban";
+		variant="m52";
+		camo="tropic";
+	};
+};
+class 34thPRC_ArmourStandard_ODST_M52_Rucksack_Medic_Tropic : 34thPRC_ArmourStandard_ODST_M52_Rucksack_Medic_Urban
+{
+	displayName="[34th] M52-C ODST Rucksack Tropic";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\odst2\tropic\34th_odst_rucksack_medic_tropic.paa",
+		"MA_Armor\data\Backpacks\M52_Rucksack\Straps_CO.paa"
+	};
+	class XtdGearInfo
+	{
+		model="34thPRC_ArmourStandard_ODST_M52_Rucksack_Urban";
+		variant="c";
+		camo="tropic";
+	};
+};
+class 34thPRC_ArmourStandard_ODST_M52_Rucksack_Gunbag_Tropic : 34thPRC_ArmourStandard_ODST_M52_Rucksack_Gunbag_Urban
+{
+	displayName="[34th] M52-G ODST Rucksack Tropic";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\odst2\tropic\34th_odst_rucksack_tropic.paa",
+		"MA_Armor\data\Backpacks\M52_Rucksack\Straps_CO.paa"
+	};
+	class XtdGearInfo
+	{
+		model="34thPRC_ArmourStandard_ODST_M52_Rucksack_Urban";
+		variant="g";
+		camo="tropic";
+	};
+};
+class 34thPRC_ArmourStandard_ODST_M52_Rucksack_Radio_Tropic : 34thPRC_ArmourStandard_ODST_M52_Rucksack_Radio_Urban
+{
+	displayName="[34th] M52-R ODST Rucksack Tropic";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\odst2\tropic\34th_odst_rucksack_tropic.paa",
+		"MA_Armor\data\Backpacks\M52_Rucksack\Straps_CO.paa"
+	};
+		class XtdGearInfo
+	{
+		model="34thPRC_ArmourStandard_ODST_M52_Rucksack_Urban";
+		variant="r";
+		camo="tropic";
+	};
+};