+++ src/Addons/34thPRC_ArmourStandard/data/odst2/woodland/config_vehicles.hpp
@@ -1,0 +1,60 @@
+class 34thPRC_ArmourStandard_ODST_M52_Rucksack_Woodland : 34thPRC_ArmourStandard_ODST_M52_Rucksack_Urban
+{
+	displayName="[34th] M52 ODST Rucksack Woodland";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\odst2\woodland\34th_odst_rucksack_woodland.paa",
+		"MA_Armor\data\Backpacks\M52_Rucksack\Straps_CO.paa"
+	};
+	class XtdGearInfo
+	{
+		model="34thPRC_ArmourStandard_ODST_M52_Rucksack_Urban";
+		variant="m52";
+		camo="woodland";
+	};
+};
+class 34thPRC_ArmourStandard_ODST_M52_Rucksack_Medic_Woodland : 34thPRC_ArmourStandard_ODST_M52_Rucksack_Medic_Urban
+{
+	displayName="[34th] M52-C ODST Rucksack Woodland";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\odst2\woodland\34th_odst_rucksack_medic_woodland.paa",
+		"MA_Armor\data\Backpacks\M52_Rucksack\Straps_CO.paa"
+	};
+	class XtdGearInfo
+	{
+		model="34thPRC_ArmourStandard_ODST_M52_Rucksack_Urban";
+		variant="c";
+		camo="woodland";
+	};
+};
+class 34thPRC_ArmourStandard_ODST_M52_Rucksack_Gunbag_Woodland : 34thPRC_ArmourStandard_ODST_M52_Rucksack_Gunbag_Urban
+{
+	displayName="[34th] M52-G ODST Rucksack Woodland";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\odst2\woodland\34th_odst_rucksack_woodland.paa",
+		"MA_Armor\data\Backpacks\M52_Rucksack\Straps_CO.paa"
+	};
+	class XtdGearInfo
+	{
+		model="34thPRC_ArmourStandard_ODST_M52_Rucksack_Urban";
+		variant="g";
+		camo="woodland";
+	};
+};
+class 34thPRC_ArmourStandard_ODST_M52_Rucksack_Radio_Woodland : 34thPRC_ArmourStandard_ODST_M52_Rucksack_Radio_Urban
+{
+	displayName="[34th] M52-R ODST Rucksack Woodland";
+	hiddenSelectionsTextures[]=
+	{
+		"34thPRC_ArmourStandard\data\odst2\woodland\34th_odst_rucksack_woodland.paa",
+		"MA_Armor\data\Backpacks\M52_Rucksack\Straps_CO.paa"
+	};
+		class XtdGearInfo
+	{
+		model="34thPRC_ArmourStandard_ODST_M52_Rucksack_Urban";
+		variant="r";
+		camo="woodland";
+	};
+};