+++ src/changelog.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### New
+- M52 ODST Rucksack and variants (corpsman, gunbag, rto/jtac), and rvmat.
 
 ## 0.24.3
 ### Updated